divided the connection.py
