+++ attention/ImportanceDiffusionAgent/fluidDiffusion/__init__.py
@@ -1,0 +1,6 @@
+import sys
+import os
+
+_pkg_dir = os.path.dirname(os.path.abspath(__file__))
+if _pkg_dir not in sys.path:
+    sys.path.insert(0, _pkg_dir)
