+++ attention/ImportanceDiffusionAgent/fluidDiffusion/fluid_integration.metta
@@ -20,7 +20,7 @@
     (let*
         (
             ($pairs (extractStiPairs (getAfAtoms)))
-            ($newStis (py-call (connection.fluid_from_af $graphFile $pairs 70 70 0.9 ("insecticide" "aflatoxin"))))
+            ($newStis (py-call (connection.fluid_from_af $graphFile $pairs 70 70 0.1 ("abamectin" "beanbug" "beetle") 0.2 0.1 "value_alignment" True False)))
         )
         (applyNewStis $newStis)
     )
